requirements.txt
playwright
google-genai
python-dotenv
spotipy
fastapi
uvicorn[standard]
pydantic
requests
aiohttp
